requirements.txt
Flask==2.3.3
pandas==2.2.2
openpyxl==3.1.2
gunicorn==21.2.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.100.0
werkzeug==2.3.8
firebase-admin==6.5.0
requests==2.31.0
httplib2==0.22.0
google-auth-httplib2==0.2.0
pyrebase4==4.7.1
